refactor(channel_finder): split feedback view into focused modules
Break the feedback view into a lifecycle+list core plus a detail module,
pure render/parse helpers, and a shared view-state module, keeping the
mountFeedback/unmountFeedback surface unchanged. The list<->detail re-render is
decoupled through registered hooks rather than a circular import. Adds unit
tests for the extracted helpers.
