+++ src/osprey/interfaces/channel_finder/static/js/feedback.js
@@ -1,38 +1,51 @@
-// @ts-nocheck
-// TODO(frontend-hardening Pn): remove & fix types when this interface is retrofitted (P2–P5)
+// @ts-check
 /**
- * OSPREY Channel Finder — Feedback Management View
+ * OSPREY Channel Finder — Feedback Management View (lifecycle + list).
  *
  * Browse and manage feedback entries (successful/failed navigation paths)
- * used by the hierarchical pipeline for search hints.
+ * used by the hierarchical pipeline for search hints. Detail-view rendering and
+ * per-record mutations live in feedback-detail.js; pure string/parse helpers in
+ * feedback-render.js; shared view-state in feedback-state.js.
  */
 
-import { fetchJSON, postJSON, putJSON, deleteJSON } from './api.js';
-import { esc } from './utils.js';
+import { fetchJSON, postJSON, deleteJSON } from './api.js';
+import { esc, messageOf } from './utils.js';
 import { formModal, confirmModal } from './modal.js';
 import { showToast } from './app.js';
-
-let _container = null;
-let _currentKey = null; // When set, we're in detail view
+import { getContainer, setContainer, getCurrentKey, setCurrentKey, setRerender, setRenderList } from './feedback-state.js';
+import { _renderDetail } from './feedback-detail.js';
+import {
+  _toolLabel,
+  _buildCardSummary,
+  _parseChannelsFromResponse,
+  _renderChannelList,
+  _renderSelections,
+  _parseSelections,
+  _formatTime,
+} from './feedback-render.js';
 
 // ---- Public API ----
 
+/**
+ * @param {HTMLElement} container
+ */
 export function mountFeedback(container) {
-  _container = container;
-  _currentKey = null;
+  setContainer(container);
+  setCurrentKey(null);
   _render();
 }
 
 export function unmountFeedback() {
-  _container = null;
-  _currentKey = null;
+  setContainer(null);
+  setCurrentKey(null);
 }
 
 // ---- Rendering ----
 
 async function _render() {
-  if (!_container) return;
-  _container.innerHTML = '<div class="loading-center"><div class="loading-spinner"></div> Loading feedback...</div>';
+  const container = getContainer();
+  if (!container) return;
+  container.innerHTML = '<div class="loading-center"><div class="loading-spinner"></div> Loading feedback...</div>';
 
   try {
     const status = await fetchJSON('/api/feedback/status');
@@ -40,18 +53,20 @@ async function _render() {
       _renderDisabled();
       return;
     }
-    if (_currentKey) {
+    if (getCurrentKey()) {
       await _renderDetail();
     } else {
       await _renderList();
     }
   } catch (e) {
-    _container.innerHTML = `<div class="empty-state"><div class="empty-state-icon">!</div>Failed to load feedback: ${esc(e.message)}</div>`;
+    container.innerHTML = `<div class="empty-state"><div class="empty-state-icon">!</div>Failed to load feedback: ${esc(messageOf(e))}</div>`;
   }
 }
 
 function _renderDisabled() {
-  _container.innerHTML = `
+  const container = getContainer();
+  if (!container) return;
+  container.innerHTML = `
     <div class="fb-disabled">
       <div class="fb-disabled-icon">&#128274;</div>
       <div class="fb-disabled-title">Feedback Store Not Available</div>
@@ -66,17 +81,20 @@ function _renderDisabled() {
 // ---- List View ----
 
 async function _renderList() {
+  const container = getContainer();
+  if (!container) return;
+
   const [data, pendingData] = await Promise.all([
     fetchJSON('/api/feedback'),
     fetchJSON('/api/pending-reviews/status')
-      .then(s => s.available ? fetchJSON('/api/pending-reviews') : { items: [] })
+      .then((/** @type {any} */ s) => s.available ? fetchJSON('/api/pending-reviews') : { items: [] })
       .catch(() => ({ items: [] })),
   ]);
   const entries = data.entries || [];
   const pendingItems = pendingData.items || [];
 
-  const totalSuccesses = entries.reduce((s, e) => s + e.success_count, 0);
-  const totalFailures = entries.reduce((s, e) => s + e.failure_count, 0);
+  const totalSuccesses = entries.reduce((/** @type {number} */ s, /** @type {any} */ e) => s + e.success_count, 0);
+  const totalFailures = entries.reduce((/** @type {number} */ s, /** @type {any} */ e) => s + e.failure_count, 0);
 
   let html = `
     <div class="fb-toolbar">
@@ -116,7 +134,7 @@ async function _renderList() {
             </tr>
           </thead>
           <tbody>
-            ${entries.map(e => `
+            ${entries.map((/** @type {any} */ e) => `
               <tr class="fb-table-row" data-key="${esc(e.key)}">
                 <td>${esc(e.query)}</td>
                 <td><span class="font-mono" style="font-size: var(--text-xs);">${esc(e.facility)}</span></td>
@@ -142,7 +160,7 @@ async function _renderList() {
           <span>Pending Reviews <span class="fb-pending-badge">${pendingItems.length}</span></span>
           <span class="fb-pending-subtitle">Agent-captured searches awaiting review</span>
         </div>
-        ${pendingItems.map(item => {
+        ${pendingItems.map((/** @type {any} */ item) => {
           const summary = _buildCardSummary(item);
           const label = _toolLabel(item.tool_name);
           const channels = _parseChannelsFromResponse(item.tool_response);
@@ -189,26 +207,29 @@ async function _renderList() {
     `;
   }
 
-  _container.innerHTML = html;
+  container.innerHTML = html;
   _bindListEvents();
   _bindPendingEvents();
 }
 
 function _bindListEvents() {
+  const container = getContainer();
+  if (!container) return;
+
   // Row click → detail view
-  _container.querySelectorAll('.fb-table-row').forEach(row => {
+  container.querySelectorAll('.fb-table-row').forEach(row => {
     row.addEventListener('click', (e) => {
-      if (e.target.closest('.fb-row-delete')) return;
-      _currentKey = row.dataset.key;
+      if (/** @type {HTMLElement} */ (e.target).closest('.fb-row-delete')) return;
+      setCurrentKey(/** @type {HTMLElement} */ (row).dataset.key ?? null);
       _render();
     });
   });
 
   // Delete entry buttons
-  _container.querySelectorAll('.fb-row-delete').forEach(btn => {
+  container.querySelectorAll('.fb-row-delete').forEach(btn => {
     btn.addEventListener('click', async (e) => {
       e.stopPropagation();
-      const key = btn.dataset.key;
+      const key = /** @type {HTMLElement} */ (btn).dataset.key;
       const ok = await confirmModal({
         title: 'Delete Feedback Entry',
         message: 'This will remove the entire feedback entry and all its records.',
@@ -221,44 +242,47 @@ function _bindListEvents() {
           showToast('Entry deleted', 'success');
           _render();
         } catch (err) {
-          showToast(`Delete failed: ${err.message}`, 'error');
+          showToast(`Delete failed: ${messageOf(err)}`, 'error');
         }
       }
     });
   });
 
   // Add entry
-  const addBtn = _container.querySelector('#fb-add-btn');
+  const addBtn = container.querySelector('#fb-add-btn');
   if (addBtn) addBtn.addEventListener('click', _handleAddEntry);
 
   // Export
-  const exportBtn = _container.querySelector('#fb-export-btn');
+  const exportBtn = container.querySelector('#fb-export-btn');
   if (exportBtn) exportBtn.addEventListener('click', _handleExport);
 
   // Clear all
-  const clearBtn = _container.querySelector('#fb-clear-btn');
+  const clearBtn = container.querySelector('#fb-clear-btn');
   if (clearBtn) clearBtn.addEventListener('click', _handleClearAll);
 }
 
 function _bindPendingEvents() {
+  const container = getContainer();
+  if (!container) return;
+
   // Approve buttons — directly promote to feedback store (no modal)
-  _container.querySelectorAll('.fb-pending-approve').forEach(btn => {
+  container.querySelectorAll('.fb-pending-approve').forEach(btn => {
     btn.addEventListener('click', async () => {
-      const id = btn.dataset.id;
+      const id = /** @type {HTMLElement} */ (btn).dataset.id;
       try {
         await postJSON(`/api/pending-reviews/${id}/approve`);
         showToast('Review approved and recorded', 'success');
         _render();
       } catch (err) {
-        showToast(`Approve failed: ${err.message}`, 'error');
+        showToast(`Approve failed: ${messageOf(err)}`, 'error');
       }
     });
   });
 
   // Dismiss buttons
-  _container.querySelectorAll('.fb-pending-dismiss').forEach(btn => {
+  container.querySelectorAll('.fb-pending-dismiss').forEach(btn => {
     btn.addEventListener('click', async () => {
-      const id = btn.dataset.id;
+      const id = /** @type {HTMLElement} */ (btn).dataset.id;
       const ok = await confirmModal({
         title: 'Dismiss Pending Review',
         message: 'This will discard the captured search result without recording feedback.',
@@ -272,243 +296,20 @@ function _bindPendingEvents() {
         showToast('Review dismissed', 'success');
         _render();
       } catch (err) {
-        showToast(`Dismiss failed: ${err.message}`, 'error');
+        showToast(`Dismiss failed: ${messageOf(err)}`, 'error');
       }
     });
   });
 
   // Expand/collapse channel list toggles
-  _container.querySelectorAll('.fb-pending-channels-toggle').forEach(btn => {
+  container.querySelectorAll('.fb-pending-channels-toggle').forEach(btn => {
     btn.addEventListener('click', () => {
-      const overflow = btn.previousElementSibling;
+      const overflow = /** @type {HTMLElement|null} */ (btn.previousElementSibling);
       if (!overflow) return;
       const isHidden = overflow.style.display === 'none';
       overflow.style.display = isHidden ? 'contents' : 'none';
-      const count = btn.dataset.full;
+      const count = /** @type {HTMLElement} */ (btn).dataset.full;
       btn.textContent = isHidden ? 'show less' : `+${count} more`;
-    });
-  });
-}
-
-// ---- Detail View ----
-
-async function _renderDetail() {
-  let entry;
-  try {
-    entry = await fetchJSON(`/api/feedback/${_currentKey}`);
-  } catch (e) {
-    showToast('Entry not found', 'error');
-    _currentKey = null;
-    await _renderList();
-    return;
-  }
-
-  const meta = entry._meta || {};
-  const successes = entry.successes || [];
-  const failures = entry.failures || [];
-
-  const html = `
-    <div class="fb-breadcrumb">
-      <a id="fb-back-link">Feedback</a>
-      <span class="fb-breadcrumb-sep">/</span>
-      <span class="fb-breadcrumb-current">"${esc(meta.query || '(unknown)')}" (${esc(meta.facility || '?')})</span>
-    </div>
-
-    <div class="fb-detail-grid">
-      <div class="fb-detail-section">
-        <div class="fb-detail-section-header fb-success-header">
-          <span>Successes (${successes.length})</span>
-          <button class="btn btn-sm btn-secondary" id="fb-add-success">+ Add</button>
-        </div>
-        ${successes.length === 0 ? '<div class="empty-state" style="padding: var(--space-4);">No successes</div>' : ''}
-        ${successes.map((s, i) => _renderSuccessCard(s, i)).join('')}
-      </div>
-
-      <div class="fb-detail-section">
-        <div class="fb-detail-section-header fb-failure-header">
-          <span>Failures (${failures.length})</span>
-          <button class="btn btn-sm btn-secondary" id="fb-add-failure">+ Add</button>
-        </div>
-        ${failures.length === 0 ? '<div class="empty-state" style="padding: var(--space-4);">No failures</div>' : ''}
-        ${failures.map((f, i) => _renderFailureCard(f, i)).join('')}
-      </div>
-    </div>
-  `;
-
-  _container.innerHTML = html;
-  _bindDetailEvents(meta);
-}
-
-function _renderSuccessCard(record, index) {
-  return `
-    <div class="fb-record-card" data-type="successes" data-index="${index}">
-      <div class="fb-selections">${_renderSelections(record.selections)}</div>
-      <div class="fb-record-meta">
-        <span class="fb-channel-count">${record.channel_count} channels</span>
-        <span class="fb-timestamp">${_formatTime(record.timestamp)}</span>
-        <span class="fb-record-actions">
-          <button class="item-action-btn action-edit fb-edit-record" title="Edit">&#9998;</button>
-          <button class="item-action-btn action-delete fb-delete-record" title="Delete">&times;</button>
-        </span>
-      </div>
-    </div>
-  `;
-}
-
-function _renderFailureCard(record, index) {
-  return `
-    <div class="fb-record-card" data-type="failures" data-index="${index}">
-      <div class="fb-selections">${_renderSelections(record.partial_selections)}</div>
-      ${record.reason ? `<div class="fb-record-reason">${esc(record.reason)}</div>` : ''}
-      <div class="fb-record-meta">
-        <span class="fb-timestamp">${_formatTime(record.timestamp)}</span>
-        <span class="fb-record-actions">
-          <button class="item-action-btn action-edit fb-edit-record" title="Edit">&#9998;</button>
-          <button class="item-action-btn action-delete fb-delete-record" title="Delete">&times;</button>
-        </span>
-      </div>
-    </div>
-  `;
-}
-
-function _toolLabel(toolName) {
-  if (!toolName) return '';
-  if (toolName.includes('build_channels')) return 'build';
-  if (toolName.includes('get_channels')) return 'get';
-  return '';
-}
-
-function _buildSelectionPath(selections) {
-  if (!selections || Object.keys(selections).length === 0) return '';
-  const order = ['system', 'family', 'device', 'field', 'subfield'];
-  const parts = [];
-  const used = new Set();
-
-  for (const key of order) {
-    if (!(key in selections)) continue;
-    used.add(key);
-    const val = selections[key];
-    if (key === 'subfield') {
-      // Merge with previous field part if field was present
-      if (parts.length > 0 && 'field' in selections) {
-        parts[parts.length - 1] += ':' + (Array.isArray(val) ? val.join(', ') : String(val));
-        continue;
-      }
-    }
-    if (Array.isArray(val)) {
-      if (val.length <= 2) {
-        parts.push(val.join(', '));
-      } else {
-        parts.push(`${val[0]}, ${val[1]} +${val.length - 2}`);
-      }
-    } else {
-      parts.push(String(val));
-    }
-  }
-
-  // Append non-standard keys
-  for (const [k, v] of Object.entries(selections)) {
-    if (used.has(k)) continue;
-    const vs = Array.isArray(v) ? v.join(', ') : String(v);
-    parts.push(`${k}:${vs}`);
-  }
-
-  return parts.join(' / ');
-}
-
-function _buildCardSummary(item) {
-  if (item.query && item.query.trim()) return item.query.trim();
-  const path = _buildSelectionPath(item.selections);
-  if (path) return path;
-  const label = _toolLabel(item.tool_name);
-  if (label) return label;
-  return 'Agent-captured search';
-}
-
-function _parseChannelsFromResponse(toolResponse) {
-  if (!toolResponse) return [];
-  try {
-    let parsed = typeof toolResponse === 'string' ? JSON.parse(toolResponse) : toolResponse;
-    // Double-encoded: {result: "..."} envelope
-    if (parsed.result && typeof parsed.result === 'string') {
-      try { parsed = JSON.parse(parsed.result); } catch { /* use as-is */ }
-    }
-    const channels = parsed.channels;
-    if (!Array.isArray(channels)) return [];
-    return channels.map(ch => {
-      if (typeof ch === 'string') return ch;
-      if (ch && typeof ch === 'object') return ch.name || ch.pv || JSON.stringify(ch);
-      return String(ch);
-    });
-  } catch {
-    return [];
-  }
-}
-
-function _renderChannelList(channels, visibleCount = 3) {
-  if (!channels || channels.length === 0) return '';
-  const visible = channels.slice(0, visibleCount);
-  const overflow = channels.slice(visibleCount);
-  let html = '<div class="fb-pending-channels">';
-  html += '<div class="fb-pending-channels-label">Channels</div>';
-  html += '<div class="fb-pending-channels-list">';
-  html += visible.map(ch => `<span class="fb-pending-pv">${esc(ch)}</span>`).join('');
-  if (overflow.length > 0) {
-    html += '<div class="fb-pending-channels-overflow" style="display:none">';
-    html += overflow.map(ch => `<span class="fb-pending-pv">${esc(ch)}</span>`).join('');
-    html += '</div>';
-    html += `<button type="button" class="fb-pending-channels-toggle" data-full="${overflow.length}">+${overflow.length} more</button>`;
-  }
-  html += '</div></div>';
-  return html;
-}
-
-function _renderSelections(selections) {
-  if (!selections || Object.keys(selections).length === 0) {
-    return '<span style="color: var(--text-muted); font-size: var(--text-xs);">(empty)</span>';
-  }
-  return Object.entries(selections).map(([k, v]) => {
-    const val = Array.isArray(v) ? v.join(', ') : String(v);
-    return `<span class="fb-sel-pair"><span class="fb-sel-key">${esc(k)}</span><span class="fb-sel-value">${esc(val)}</span></span>`;
-  }).join('');
-}
-
-function _bindDetailEvents(meta) {
-  // Back link
-  _container.querySelector('#fb-back-link')?.addEventListener('click', () => {
-    _currentKey = null;
-    _render();
-  });
-
-  // Add success to existing entry
-  _container.querySelector('#fb-add-success')?.addEventListener('click', async () => {
-    await _handleAddRecordToEntry(meta, 'success');
-  });
-
-  // Add failure to existing entry
-  _container.querySelector('#fb-add-failure')?.addEventListener('click', async () => {
-    await _handleAddRecordToEntry(meta, 'failure');
-  });
-
-  // Edit record buttons
-  _container.querySelectorAll('.fb-edit-record').forEach(btn => {
-    btn.addEventListener('click', async (e) => {
-      e.stopPropagation();
-      const card = btn.closest('.fb-record-card');
-      const type = card.dataset.type;
-      const index = parseInt(card.dataset.index, 10);
-      await _handleEditRecord(type, index);
-    });
-  });
-
-  // Delete record buttons
-  _container.querySelectorAll('.fb-delete-record').forEach(btn => {
-    btn.addEventListener('click', async (e) => {
-      e.stopPropagation();
-      const card = btn.closest('.fb-record-card');
-      const type = card.dataset.type;
-      const index = parseInt(card.dataset.index, 10);
-      await _handleDeleteRecord(type, index);
     });
   });
 }
@@ -545,145 +346,7 @@ async function _handleAddEntry() {
     showToast('Entry added', 'success');
     _render();
   } catch (err) {
-    showToast(`Add failed: ${err.message}`, 'error');
-  }
-}
-
-async function _handleAddRecordToEntry(meta, entryType) {
-  const fields = [
-    { name: 'selections', label: 'Selections (key: value, one per line)', type: 'textarea', placeholder: 'system: MAG\ndevice: QF1' },
-  ];
-  if (entryType === 'success') {
-    fields.push({ name: 'channel_count', label: 'Channel Count', type: 'number', value: '0' });
-  } else {
-    fields.push({ name: 'reason', label: 'Failure Reason', type: 'text', placeholder: 'e.g. no options at family level' });
-  }
-
-  const result = await formModal({
-    title: `Add ${entryType === 'success' ? 'Success' : 'Failure'} Record`,
-    fields,
-    submitLabel: 'Add',
-  });
-
-  if (!result) return;
-
-  const selections = _parseSelections(result.selections);
-  try {
-    await postJSON('/api/feedback', {
-      query: meta.query || '',
-      facility: meta.facility || '',
-      entry_type: entryType,
-      selections,
-      channel_count: parseInt(result.channel_count || '0', 10),
-      reason: result.reason || '',
-    });
-    showToast('Record added', 'success');
-    _render();
-  } catch (err) {
-    showToast(`Add failed: ${err.message}`, 'error');
-  }
-}
-
-async function _handleEditRecord(recordType, index) {
-  // Fetch fresh entry data
-  let entry;
-  try {
-    entry = await fetchJSON(`/api/feedback/${_currentKey}`);
-  } catch (e) {
-    showToast('Entry not found', 'error');
-    return;
-  }
-
-  const records = entry[recordType] || [];
-  const record = records[index];
-  if (!record) {
-    showToast('Record not found', 'error');
-    return;
-  }
-
-  const isSuccess = recordType === 'successes';
-  const currentSelections = isSuccess ? record.selections : record.partial_selections;
-  const selStr = Object.entries(currentSelections || {}).map(([k, v]) => `${k}: ${v}`).join('\n');
-
-  const fields = [
-    { name: 'selections', label: 'Selections (key: value, one per line)', type: 'textarea', value: selStr },
-  ];
-  if (isSuccess) {
-    fields.push({ name: 'channel_count', label: 'Channel Count', type: 'number', value: String(record.channel_count || 0) });
-  } else {
-    fields.push({ name: 'reason', label: 'Failure Reason', type: 'text', value: record.reason || '' });
-  }
-
-  const result = await formModal({
-    title: `Edit ${isSuccess ? 'Success' : 'Failure'} Record`,
-    fields,
-    submitLabel: 'Save',
-  });
-
-  if (!result) return;
-
-  const selections = _parseSelections(result.selections);
-  const body = { expected_timestamp: record.timestamp };
-  if (isSuccess) {
-    body.selections = selections;
-    body.channel_count = parseInt(result.channel_count || '0', 10);
-  } else {
-    body.partial_selections = selections;
-    body.reason = result.reason || '';
-  }
-
-  try {
-    await putJSON(`/api/feedback/${_currentKey}/${recordType}/${index}`, body);
-    showToast('Record updated', 'success');
-    _render();
-  } catch (err) {
-    if (err.status === 409) {
-      showToast('Record changed by another user, refreshing...', 'info');
-      _render();
-    } else {
-      showToast(`Edit failed: ${err.message}`, 'error');
-    }
-  }
-}
-
-async function _handleDeleteRecord(recordType, index) {
-  // Fetch fresh timestamp
-  let entry;
-  try {
-    entry = await fetchJSON(`/api/feedback/${_currentKey}`);
-  } catch (e) {
-    showToast('Entry not found', 'error');
-    return;
-  }
-
-  const records = entry[recordType] || [];
-  const record = records[index];
-  if (!record) {
-    showToast('Record not found', 'error');
-    return;
-  }
-
-  const ok = await confirmModal({
-    title: 'Delete Record',
-    message: `Delete this ${recordType === 'successes' ? 'success' : 'failure'} record?`,
-    confirmLabel: 'Delete',
-    danger: true,
-  });
-  if (!ok) return;
-
-  try {
-    await deleteJSON(`/api/feedback/${_currentKey}/${recordType}/${index}`, {
-      expected_timestamp: record.timestamp,
-    });
-    showToast('Record deleted', 'success');
-    _render();
-  } catch (err) {
-    if (err.status === 409) {
-      showToast('Record changed, refreshing...', 'info');
-      _render();
-    } else {
-      showToast(`Delete failed: ${err.message}`, 'error');
-    }
+    showToast(`Add failed: ${messageOf(err)}`, 'error');
   }
 }
 
@@ -700,7 +363,7 @@ async function _handleExport() {
     URL.revokeObjectURL(url);
     showToast('Export downloaded', 'success');
   } catch (err) {
-    showToast(`Export failed: ${err.message}`, 'error');
+    showToast(`Export failed: ${messageOf(err)}`, 'error');
   }
 }
 
@@ -719,37 +382,13 @@ async function _handleClearAll() {
     showToast('All feedback cleared', 'success');
     _render();
   } catch (err) {
-    showToast(`Clear failed: ${err.message}`, 'error');
+    showToast(`Clear failed: ${messageOf(err)}`, 'error');
   }
 }
 
-// ---- Helpers ----
-
-function _parseSelections(text) {
-  const selections = {};
-  if (!text) return selections;
-  for (const line of text.split('\n')) {
-    const trimmed = line.trim();
-    if (!trimmed) continue;
-    const colonIdx = trimmed.indexOf(':');
-    if (colonIdx > 0) {
-      const key = trimmed.slice(0, colonIdx).trim();
-      const value = trimmed.slice(colonIdx + 1).trim();
-      if (key) selections[key] = value;
-    }
-  }
-  return selections;
-}
-
-function _formatTime(isoStr) {
-  if (!isoStr) return '';
-  try {
-    const d = new Date(isoStr);
-    return d.toLocaleString(undefined, {
-      month: 'short', day: 'numeric',
-      hour: '2-digit', minute: '2-digit',
-    });
-  } catch {
-    return isoStr;
-  }
-}
+// Register the list/detail dispatcher and the direct list renderer so
+// feedback-detail.js can re-render without importing this module at eval time
+// (avoids an ESM cycle). setRerender = full dispatcher; setRenderList = the
+// list view directly (the detail "entry not found" path drops straight to it).
+setRerender(_render);
+setRenderList(_renderList);
